pyproject.toml
[project]
name = "build-a-multi-modal-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asttokens==2.4.1",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "colorama>=0.4.6",
    "comm==0.2.2",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "debugpy==1.8.2",
    "decorator==5.1.1",
    "exceptiongroup>=1.2.2",
    "executing==2.0.1",
    "filelock==3.17.0",
    "fire>=0.7.0",
    "fonttools==4.56.0",
    "fsspec==2025.3.0",
    "huggingface-hub==0.29.3",
    "idna==3.10",
    "importlib-metadata>=8.6.1",
    "ipykernel==6.29.5",
    "ipython==8.26.0",
    "ipython-pygments-lexers>=1.1.1",
    "jedi==0.19.1",
    "jinja2==3.1.6",
    "joblib==1.4.2",
    "jupyter-client==8.6.2",
    "jupyter-core==5.7.2",
    "kiwisolver==1.4.8",
    "markupsafe==3.0.2",
    "matplotlib==3.10.1",
    "matplotlib-inline==0.1.7",
    "mpmath==1.3.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "numpy==2.2.3",
    "packaging>=24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pickleshare>=0.7.5",
    "pillow==11.1.0",
    "platformdirs==4.2.2",
    "prompt-toolkit==3.0.47",
    "protobuf==6.30.0",
    "psutil==6.0.0",
    "pure-eval==0.2.2",
    "pydantic==2.8.2",
    "pydantic-core==2.20.1",
    "pydantic-settings==2.3.4",
    "pygments==2.18.0",
    "pyparsing==3.2.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "pywin32==306",
    "pyyaml==6.0.2",
    "pyzmq==26.0.3",
    "regex==2024.11.6",
    "requests==2.32.3",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentencepiece==0.2.0",
    "setuptools==75.8.2",
    "six>=1.17.0",
    "stack-data==0.6.3",
    "sympy==1.13.1",
    "threadpoolctl==3.5.0",
    "tokenizers==0.21.0",
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchinfo==1.8.0",
    "torchtext==0.18.0",
    "torchvision==0.21.0",
    "tornado==6.4.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.49.0",
    "typing-extensions>=4.13.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "uv==0.6.14",
    "wcwidth==0.2.13",
    "wheel==0.45.1",
    "zipp>=3.21.0",
]
